refactor: introduce subscription mod
