+++ benches/helpers.rs
@@ -132,7 +132,7 @@ pub async fn http_server(handle: tokio::runtime::Handle) -> (String, jsonrpsee::
 /// Run jsonrpsee WebSocket server for benchmarks.
 #[cfg(not(feature = "jsonrpc-crate"))]
 pub async fn ws_server(handle: tokio::runtime::Handle) -> (String, jsonrpsee::server::ServerHandle) {
-	use jsonrpsee::{core::server::rpc_module::SubscriptionMessage, server::ServerBuilder};
+	use jsonrpsee::{core::server::SubscriptionMessage, server::ServerBuilder};
 
 	let server = ServerBuilder::default()
 		.max_request_body_size(u32::MAX)
