+++ examples/examples/ws_pubsub_with_params.rs
@@ -29,7 +29,7 @@ use std::time::Duration;
 
 use futures::{Stream, StreamExt};
 use jsonrpsee::core::client::{Subscription, SubscriptionClientT};
-use jsonrpsee::core::server::rpc_module::{SubscriptionMessage, TrySendError};
+use jsonrpsee::core::server::{SubscriptionMessage, TrySendError};
 use jsonrpsee::core::Serialize;
 use jsonrpsee::server::{RpcModule, ServerBuilder};
 use jsonrpsee::types::ErrorObjectOwned;
