+++ jsonrpsee/src/lib.rs
@@ -90,7 +90,7 @@ cfg_types! {
 }
 
 cfg_server! {
-	pub use jsonrpsee_core::server::rpc_module::{RpcModule, SubscriptionSink, PendingSubscriptionSink, SubscriptionMessage, DisconnectError, TrySendError, SendTimeoutError};
+	pub use jsonrpsee_core::server::*;
 	pub use tokio;
 }
