+++ server/src/lib.rs
@@ -43,10 +43,7 @@ mod tests;
 
 pub use future::ServerHandle;
 pub use jsonrpsee_core::server::host_filtering::AllowHosts;
-pub use jsonrpsee_core::server::rpc_module::{
-	DisconnectError, PendingSubscriptionSink, RpcModule, SendTimeoutError, SubscriptionMessage, SubscriptionSink,
-	TrySendError,
-};
+pub use jsonrpsee_core::server::*;
 pub use jsonrpsee_core::{id_providers::*, traits::IdProvider};
 pub use jsonrpsee_types as types;
 pub use server::{Builder as ServerBuilder, Server};
